feat: Refactor decryption methods and remove AES dependency from WalletNode
