+++ src/wallet-node/railgun.ts
@@ -1,4 +1,7 @@
+import { AES } from '@railgun-reloaded/cryptography'
+
 import { deriveNodes } from '../index.js'
+import { getSharedSymmetricKey } from '../keys.js'
 import type {
   SpendingKeyPair,
   SpendingPublicKey,
@@ -172,5 +175,36 @@ export class RailgunWallet {
       throw new Error('Keystore not initialized')
     }
     return this.keystore.viewingKeyPair.privateKey
+  }
+
+  /**
+   * Decrypts a randomly encrypted bundle using AES-GCM.
+   * @param encryptedBundle - An array containing three Uint8Array elements:
+   *   - The first element represents the IV and authentication tag.
+   *   - The second element contains the encrypted data.
+   *   - The third element is unused in this function.
+   * @param shieldKey - A Uint8Array representing the shared key used for decryption.
+   * @returns The decrypted data as a Uint8Array.
+   * @throws Will throw an error if decryption fails or the input data is invalid.
+   */
+  decryptRandom (
+    encryptedBundle: [Uint8Array, Uint8Array, Uint8Array],
+    shieldKey: Uint8Array
+  ): Uint8Array {
+    const sharedKey = getSharedSymmetricKey(this.getViewingPublicKey(), shieldKey)
+    if (!sharedKey) {
+      throw new Error('No shared key.')
+    }
+    const hexlified0 = encryptedBundle[0]
+    const hexlified1 = encryptedBundle[1]
+    const decrypted = AES.decryptGCM(
+      {
+        iv: hexlified0.slice(0, 16),
+        tag: hexlified0.slice(16, 32),
+        data: [hexlified1.slice(0, 16)],
+      },
+      sharedKey
+    )[0]!
+    return decrypted
   }
 }
